requirements.txt
configargparse
apscheduler
https://github.com/hensur/pysmbc/archive/master.zip
python-magic